Fix empty lines in databases export
